moving default_registry into _ui_tester_register
