+++ traitsui/testing/tester/tests/test_default_registry.py
@@ -11,7 +11,7 @@
 
 import unittest
 
-from traitsui.testing.tester.default_registry import get_default_registry
+from traitsui.testing.tester._ui_tester_register.default_registry import get_default_registry
 from traitsui.testing.tester.registry import TargetRegistry
 from traitsui.tests._tools import is_null
